Fix the visits table so it has an Index
This change was made because when running the command to add more visits
it will throw an error saying that the primary keys are not unique.
This is fixed with adding an id column and adding it as a PRIMARY KEY
while removing date_of_visit from the PRIMAY KEY value.